Moved Operate Client to Starter module and cleaned up AutoConfigurations further
