+++ connector-runtime/src/main/resources/application.properties
@@ -5,6 +5,8 @@ camunda.connector.polling.interval=5000
 camunda.connector.webhook.enabled=false
 spring.main.web-application-type=none
 
+operate.client.enabled=true
+
 # Local demo installation with simple authentication, not meant for production use
 #zeebe.client.broker.gateway-address=127.0.0.1:26500
 #zeebe.client.security.plaintext=true
