+++ connector-runtime/src/test/java/io/camunda/connector/runtime/inbound/InboundConnectorTestConfiguration.java
@@ -16,30 +16,29 @@
  */
 package io.camunda.connector.runtime.inbound;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-import io.camunda.zeebe.spring.client.metrics.SimpleMetricsRecorder;
-import io.camunda.connector.runtime.inbound.operate.OperateClientLifecycle;
 import io.camunda.operate.CamundaOperateClient;
 import io.camunda.operate.exception.OperateException;
-import java.util.ArrayList;
-
 import io.camunda.zeebe.spring.client.metrics.MetricsRecorder;
+import io.camunda.zeebe.spring.client.metrics.SimpleMetricsRecorder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
+
+import java.util.ArrayList;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 @Configuration
 public class InboundConnectorTestConfiguration {
 
   @Bean
   @Primary
-  public OperateClientLifecycle operateClientLifecycle() throws OperateException {
+  public CamundaOperateClient mockOperateClient() throws OperateException {
     CamundaOperateClient camundaOperateClientMock = mock(CamundaOperateClient.class);
     when(camundaOperateClientMock.searchProcessDefinitions(any())).thenReturn(new ArrayList<>());
-    return new OperateClientLifecycle(camundaOperateClientMock);
+    return camundaOperateClientMock;
   }
 
   @Bean
