+++ spring-boot-starter-camunda/src/main/java/io/camunda/zeebe/spring/client/configuration/MetricsDefaultConfiguration.java
@@ -11,7 +11,6 @@ import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.type.AnnotatedTypeMetadata;
 
-@Configuration
 public class MetricsDefaultConfiguration {
 
   @Bean
