+++ spring-boot-starter-camunda/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.camunda.zeebe.spring.client.SpringZeebeAutoConfiguration
+io.camunda.zeebe.spring.client.CamundaAutoConfiguration
