+++ test/common/src/main/java/io/camunda/zeebe/spring/test/CamundaTestAutoConfiguration.java
@@ -1,23 +1,23 @@
 package io.camunda.zeebe.spring.test;
 
-import io.camunda.zeebe.spring.client.AbstractZeebeBaseConfiguration;
-import io.camunda.zeebe.spring.client.SpringZeebeAutoConfiguration;
+import io.camunda.zeebe.spring.client.CamundaAutoConfiguration;
 import io.camunda.zeebe.spring.client.testsupport.SpringZeebeTestContext;
 import io.camunda.zeebe.spring.test.configuration.ZeebeTestDefaultConfiguration;
 import io.camunda.zeebe.spring.test.proxy.TestProxyConfiguration;
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
-import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
 
 @TestConfiguration
-@ImportAutoConfiguration({TestProxyConfiguration.class, ZeebeTestDefaultConfiguration.class})
-@AutoConfigureBefore(SpringZeebeAutoConfiguration.class)
-@AutoConfigureAfter(JacksonAutoConfiguration.class) // make sure Spring created ObjectMapper is preferred if available
-public class ZeebeTestConfiguration extends AbstractZeebeBaseConfiguration {
+@ImportAutoConfiguration({
+  TestProxyConfiguration.class,
+  ZeebeTestDefaultConfiguration.class,
+  CamundaAutoConfiguration.class
+})
+@AutoConfigureBefore(CamundaAutoConfiguration.class)
+public class CamundaTestAutoConfiguration {
 
   @Bean
   public SpringZeebeTestContext enableTestContext() {
