+++ test/embedded/src/main/java/io/camunda/zeebe/spring/test/ZeebeSpringTest.java
@@ -1,6 +1,9 @@
 package io.camunda.zeebe.spring.test;
 
+import io.camunda.zeebe.spring.client.CamundaAutoConfiguration;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.TestExecutionListeners;
 
 import java.lang.annotation.ElementType;
@@ -15,8 +18,9 @@ import java.lang.annotation.Target;
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
-// this creates the engine and the client
-@ImportAutoConfiguration({ZeebeTestConfiguration.class})
+// this pulls in the Configuration NOT as AutoConfiguration but directly creates beans, so the marker is present
+// when the normal CamundaAutoConfiguration is used by the normal meta-inf/services  way
+@Import({CamundaTestAutoConfiguration.class})
 // this listener hooks up into test execution
 @TestExecutionListeners(listeners = ZeebeTestExecutionListener.class, mergeMode = TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS)
 public @interface ZeebeSpringTest {
